% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_tilt.R
\name{filter_tilt}
\alias{filter_tilt}
\title{Filters data form prep data function iteratively}
\usage{
filter_tilt(data, g, l)
}
\arguments{
\item{data}{tilt data frame}

\item{g}{greater than median of spline}

\item{l}{less than median of spline}
}
\value{
date from of rmssd and pnn50
}
\description{
Uses window functions native to the RHRV package and hampel window filter.
}
\examples{
filter_tilt(data, g, l)
}
