% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtilt.R
\name{ggtilt}
\alias{ggtilt}
\title{Data frame of name and content of files.}
\usage{
ggtilt(.data)
}
\arguments{
\item{.data}{data file}
}
\value{
plot
}
\description{
Read in all tilt data files in a folder at once to
be iterated through by other functions.
}
\examples{
read_tilt("path/to/folder/of/tilt/files", read_csv or ... others)
}
