% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tilt.R
\name{plot_tilt}
\alias{plot_tilt}
\title{Storing plots of RR series in folders for review and produces rmssd and pnn50 output}
\usage{
plot_tilt(.data, folder, type)
}
\arguments{
\item{.data}{dataframe}

\item{folder}{folder for the saved plots}

\item{type}{whether data are filtered or original}
}
\value{
data frame
}
\description{
Storing plots of RR series in folders for review and produces rmssd and pnn50 output
}
\examples{
plot_tilt(.data, folder, type)
}
