% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tilt.R
\name{read_tilt}
\alias{read_tilt}
\title{Data frame of name and content of files.}
\usage{
read_tilt(path, file_type)
}
\arguments{
\item{path}{string}

\item{file_type}{read function}
}
\value{
nest dataframe
}
\description{
Read in all tilt data files in a folder at once to
be iterated through by other functions.
}
\examples{
read_tilt("path/to/folder/of/tilt/files", read_csv or ... others)
}
