% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_tilt.R
\name{filter_tilt}
\alias{filter_tilt}
\title{Data frame of name and content of files.}
\usage{
filter_tilt(.data, g, l)
}
\arguments{
\item{.data}{tilt data frame}

\item{g}{greater than median of spline}

\item{l}{less than median of spline}
}
\value{
date from of rmssd and pnn50
}
\description{
Read in all tilt data files in a folder at once to
be iterated through by other functions.
}
\examples{
filter_tilt(prepdata, greaterthan, lessthan)
}
