% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tilt.R
\name{plot_tilt}
\alias{plot_tilt}
\title{Storing plots of RR series in folders for review}
\usage{
plot_tilt(.data, folder, type)
}
\arguments{
\item{.data}{dataframe}

\item{folder}{folder for the saved plots}

\item{time}{hr time var}

\item{hr}{heart beat}

\item{rr}{heart beat intervals}
}
\value{
data frame
}
\description{
Storing plots of RR series in folders for review
}
\examples{
plot_tilt(data frame, time, hr, rr, folder)
}
