% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{prep_data}
\alias{prep_data}
\title{Prep Data
This function helps to manipulate the data into a data set readable by
other tidyrhrv functions.}
\usage{
prep_data(.data, time, HR, RR)
}
\arguments{
\item{.data}{data set}

\item{time}{heart beat time interval}

\item{HR}{Heart beat series}

\item{RR}{RR series}
}
\value{
nest dataframe
}
\description{
Prep Data
This function helps to manipulate the data into a data set readable by
other tidyrhrv functions.
}
\examples{
read_tilt(.data, time, Heart beat, RR)
}
