% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{prep_data}
\alias{prep_data}
\title{Prep Data}
\usage{
prep_data(.data, time, HR, RR)
}
\arguments{
\item{time}{time variable}

\item{HR}{Heart Beat}

\item{RR}{RR series}

\item{data}{data set}
}
\value{
nest dataframe
}
\description{
Prep Data
}
\examples{
read_tilt(data, time, Heart beat, RR)
}
